Use path resolution on query args to avoid mismatching
This implementation is a lot more robust than simply hoping that we
always traversed the args tree in the same order. Args should now be
guaranteed to be bound to the correct param, as long as both sides (SQL
gen and arg assignment) have the right logic for generating paths.