Rework optionals
When an optional type is passed as a param (e.g. to `where()`), use
PostgreSQL syntax `IS NOT DISTINCT FROM` to allow passing either a value
or NULL. This allows us to generate the SQL at comptime (since we don't
know if the value is `null` until runtime).

Revert changes for u16/u32/u64 primary keys - `SERIAL`, `SMALLSERIAL`
and `BIGSERIAL` do not map to these types as they are just i16/i32/i64
that can't be negative. Using signed types makes things a lot simpler
with pg.zig.